requirements.txt
Django==5.1.6
asgiref==3.8.1
daphne==4.1.2
channels==4.2.0
channels-redis==4.2.1
psycopg==3.2.5
psycopg-binary==3.2.5
redis==5.2.1
django-tailwind==3.8.0
django-allauth==65.4.1
celery==5.4.0
django-storages==1.14.4
cloudinary==1.42.2
django-cloudinary-storage==0.3.0
whitenoise==6.9.0
gunicorn==23.0.0
python-decouple==3.8
dj-database-url==2.3.0
Pillow==11.1.0
sentry-sdk==2.22.0
django-csp==3.8
twilio>=8.0.0
reportlab==4.1.0
openpyxl==3.1.5
